int masses in Products

+++ 3.1/src/main/java/products/Packed.java
@@ -4,6 +4,6 @@ package products;
 public interface Packed {
     Packaging getPackaging();
     
-    double getNetMass();
-    double getGrossMass();
+    int getNetMass();
+    int getGrossMass();
 }
